Hide the "live" class
